third_party/WebKit/LayoutTests/inspector/sources/debugger/debugger-uncaught-promise-on-dom-exception.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/console-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script>

function testFunction()
{
    console.clear();
    debugger;
}

function runPromises()
{
    var reject;
    var p = new Promise(function(res, rej) {
        reject = rej;
    });
    p.catch(function p_catch() {
        throwDOMException();
    });
    reject(new Error("FAIL: Should not be printed to console"));
}

function throwDOMException()
{
    var a = document.createElement("div");
    var b = document.createElement("div");
    a.removeChild(b);
}

function runPromisesFromInspector()
{
    // setTimeout to cut off VM call frames from the stack trace.
    setTimeout(function timeout() {
        runPromises()
    }, 0);
}

function test()
{
    InspectorTest.setQuiet(true);
    InspectorTest.startDebuggerTest(step1);

    function step1()
    {
        InspectorTest.addConsoleViewSniffer(addMessage, true);
        InspectorTest.runTestFunctionAndWaitUntilPaused(didPause);
    }

    function didPause(callFrames, reason, breakpointIds, asyncStackTrace)
    {
        InspectorTest.evaluateInPage("runPromisesFromInspector()", resumeExecution);
    }

    function resumeExecution()
    {
        InspectorTest.resumeExecution();
    }

    function addMessage(uiMessage)
    {
        InspectorTest.expandConsoleMessages(dump);
    }

    function dump()
    {
        InspectorTest.dumpConsoleMessages(false, false, InspectorTest.textContentWithLineBreaks);
        InspectorTest.completeTest();
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests unhandled promise that was rejected with a DOM exception.
</p>

</body>
</html>
