third_party/WebKit/LayoutTests/inspector/sources/debugger/debugger-uncaught-promise-on-dom-exception-expected.txt

CONSOLE ERROR: line 30: Unhandled promise rejection
Tests unhandled promise that was rejected with a DOM exception.

Set timer for test function.
debugger-uncaught-promise-on-dom-exception.html:10 Console was cleared

    debugger-uncaught-promise-on-dom-exception.html:30 Unhandled promise rejection Promise {[[PromiseStatus]]: "rejected", [[PromiseValue]]: DOMException: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of th…}
        __proto__: Promise
        [[PromiseStatus]]: "rejected"
        [[PromiseValue]]: DOMException: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node.
        debugger-uncaught-promise-on-dom-exception.html:30 throwDOMException
        debugger-uncaught-promise-on-dom-exception.html:21 p_catch
